Put both $JVM_DIR and $JVM_DIR/server on LD_LIBRARY_PATH
This is all for JPL on Linux which needs to dynamically link
libjvm.so and libjava.so.

Problem is that on some platforms, these libs can sometimes live
in different directories.

We change the nimrodel-specific convention of JVM_DIR to point
to their common parent and try including that and a server subdir,
one or the other which may not actually exist depending on where
we are